Moving terminal_tools
