+++ mango_tango_cib/analyzers/example/example_base/main.py
@@ -1,7 +1,7 @@
 import polars as pl
 
 from analyzer_interface.context import PrimaryAnalyzerContext
-from terminal_tools import ProgressReporter
+from mango_tango_cib.terminal_tools import ProgressReporter
 
 
 def main(context: PrimaryAnalyzerContext):
