+++ mango_tango_cib/app/project_context.py
@@ -4,8 +4,8 @@ import polars as pl
 from pydantic import BaseModel
 
 from analyzer_interface import UserInputColumn as BaseUserInputColumn
-from preprocessing.series_semantic import SeriesSemantic, infer_series_semantic
 from mango_tango_cib.storage import AnalysisModel, ProjectModel
+from preprocessing.series_semantic import SeriesSemantic, infer_series_semantic
 
 from .app_context import AppContext
