+++ mango_tango_cib/components/analysis_web_server.py
@@ -5,13 +5,13 @@ from pathlib import Path
 
 from dash import Dash
 from flask import Flask, render_template
+from storage import AnalysisModel, Storage
+from terminal_tools import wait_for_key
+from terminal_tools.inception import TerminalContext
 from waitress import serve
 
 from analyzer_interface.suite import AnalyzerSuite
 from context import WebPresenterContext
-from storage import AnalysisModel, Storage
-from terminal_tools import wait_for_key
-from terminal_tools.inception import TerminalContext
 
 
 def analysis_web_server(
