+++ mango_tango_cib/components/new_analysis.py
@@ -11,7 +11,12 @@ from analyzer_interface import (
     get_data_type_compatibility_score,
 )
 from mango_tango_cib.app import ProjectContext
-from terminal_tools import draw_box, print_ascii_table, prompts, wait_for_key
+from mango_tango_cib.terminal_tools import (
+    draw_box,
+    print_ascii_table,
+    prompts,
+    wait_for_key,
+)
 
 from .context import ViewContext
 from .export_outputs import export_format_prompt, export_outputs_sequence
