+++ mango_tango_cib/components/new_project.py
@@ -3,8 +3,8 @@ from traceback import format_exc
 from typing import Optional
 
 from importing import Importer, ImporterSession, importers
-from terminal_tools import draw_box, prompts, wait_for_key
-from terminal_tools.inception import Scope
+from mango_tango_cib.terminal_tools import draw_box, prompts, wait_for_key
+from mango_tango_cib.terminal_tools.inception import Scope
 
 from .context import ViewContext
