+++ mango_tango_cib/components/project_main.py
@@ -1,7 +1,7 @@
 from colorama import Fore
 
 from mango_tango_cib.app import ProjectContext
-from terminal_tools import draw_box, prompts, wait_for_key
+from mango_tango_cib.terminal_tools import draw_box, prompts, wait_for_key
 
 from .analysis_main import analysis_main
 from .context import ViewContext
