+++ mango_tango_cib/components/select_project.py
@@ -1,7 +1,12 @@
 from typing import Optional
 
 from mango_tango_cib.app import ProjectContext
-from terminal_tools import draw_box, print_ascii_table, prompts, wait_for_key
+from mango_tango_cib.terminal_tools import (
+    draw_box,
+    print_ascii_table,
+    prompts,
+    wait_for_key,
+)
 
 from .context import ViewContext
